etl/requirements.txt
pika
python-dotenv
psycopg2-binary
requests